Fix stack corruption

+++ src/depressed_cubic.cpp
@@ -11,26 +11,26 @@ DepressedCubic::DepressedCubic(float a, float b, float c, float d) {
 
 float DepressedCubic::single_root() {
 	if (discriminant > 0.0f) {
-		const float D = sqrt(q * q / 4.0f + p * p * p / 27.0f);
+		const float D = sqrtf(q * q / 4.0f + p * p * p / 27.0f);
 		const float C0 = -0.5f * q + D;
 		const float C1 = -0.5f * q - D;
 
-		root = sign(C0) * pow(abs(C0), 1.0f / 3.0f) + sign(C1) * pow(abs(C1), 1.0f / 3.0f);
+		root = sign(C0) * powf(abs(C0), 1.0f / 3.0f) + sign(C1) * powf(abs(C1), 1.0f / 3.0f);
 	}
 	return root;
 }
 
 float DepressedCubic::first_root() {
-	if (discriminant <= 0.0f) root = fac * cos(arccos);
+	if (discriminant <= 0.0f) root = fac * cosf(arccos);
 	return root;
 }
 
 float DepressedCubic::second_root() {
-	if (discriminant <= 0.0f) root = fac * cos(arccos - 2.0f / 3.0f * M_PI);
+	if (discriminant <= 0.0f) root = fac * cosf(arccos - 2.0f / 3.0f * (float) M_PI);
 	return root;
 }
 
 float DepressedCubic::third_root() {
-	if (discriminant <= 0.0f) root = fac * cos(arccos - 4.0f / 3.0f * M_PI);
+	if (discriminant <= 0.0f) root = fac * cosf(arccos - 4.0f / 3.0f * (float) M_PI);
 	return root;
 }
